move trace_block_with and trace_block_until to EthTransactions
